Preliminary implementation of showing and hiding properties through validateConfig. There are two extra groups of properties, one group for HTTP tunneling and another for separate authentication. These can be expanded or hidden (and disabled) through the GSA admin console.
Also some small cleanup of the properties, including a standard 
ordering in the code and config files, and renaming the *caRootCerts 
properties to *caRootCert, since they are just single certificates. 
The XML/LivelinkConnector property names are still *caRootCerts, 
because those are lists, using the *caRootCert properties for one 
element of each list.