Refactored the AuthenticationManager creation as part of adding support for an LDAP authentication manager.
The AuthenticationManager is now configured as a bean attribute
of the connector. The default implementation remains
LivelinkAuthenticationManager. The connector is responsible for
checking to see if the AuthenticationManager implements the new
ConnectorAware interface and passing itself to the
AuthenticationManager if so to allow the AuthenticationManager to
finish configuring itself.

A new AuthenticationManager, AuthenticationManagerChain, was
added. This class wraps a list of one or more
AuthenticationManagers and tries each in turn. 

A new AuthenticationManager, LdapAuthenticationManager, was
added. This AuthenticationManager authenticates against an LDAP
server rather than against Livelink. This AuthenticationManager
must be configured in connectorInstance.xml. This
AuthenticationManager may be enhanced to accept more
configuration parameters later. It has also been proposed that
this, along with AuthenticationManagerChain, be made part of the
Connector Manager for use by other connectors.