Moved the messages file from com/google/enterprise/connector/otex/LivelinkConnectorType.properties to config/OtexConnectorResources_en.properties. This is the standard location/naming convention for connector resource files.
Provided message file entries for exceptions thrown in
LivelinkConnectorType and LivelinkConnector if they can get back
to the user on the config form. Used getLocalizedMessage when
getting messages from exceptions; it's mostly not implemented but
defaults to getMessage. 

Added try/catch blocks to methods in LivelinkConnectorType so
that no exceptions are thrown. Errors are always reported to the
user either in the message part of the ConfigureResponse or in
the form snippet if needed. getConfigForm, for example, seems to
ignore any message in the ConfigureResponse.

Added code to LivelinkException to support supplying a message
key and optional parameters so that exceptions which might make
it to the user can be translated.