Require a non-empty server value to avoid the LAPI default of localhost.
If NTLM is enabled, require an HTTP username and password to
avoid the LAPI default of grabbing the current process's
information.

Remove enableNtlm from the form and set it to false by default in
connectorInstance.xml. It hasn't been tested and requires extra
setup, so we want to make it harder to do.