Changed the "password" property to "Password" so that it would be encrypted in the connector instance properties files. This depends on Connector Manager revision 684 to work properly. The renaming was just in the form entry in LivelinkConnectorType and the placeholder in connectorInstance.xml.
Existing connector instances will either need to be deleted and
recreated, or carefully manually updated:

1. Edit the properties file and change the password property
   to Password with a blank value ("Password=").
2. Update the Connector Manager and Livelink connector JAR files.
3. Restart Tomcat.
4. Edit the connector in the admin console to restore the password.

Step 1 cannot be performed in the admin console because we need to
change the property name. We need to delete the password value so that
it can be decrypted successfully. Step 4 cannot be performed by
editing the properties file because we need the encrypted password
value.

Small code cleanup (unused imports, comment tweak, indentation) in
LivelinkConnectorType and LivelinkConnector.