Removed the following properties from the configuration form: connection, httpUsername, httpPassword, verifyServer, caRootCert, authenticationConnection, authenticationVerifyServer, and authenticationCaRootCert.
Moved username, password, and domainName into a new always-visible
Livelink System Administrator group, presaging the changes to come.

Removed the group indentation, just because I thought it looked better
with the new group.

Marked some of the tunneling and separate authentication properties as
"required".

Added a comment about the sketchy use of
PropertyPlaceholderConfigurer.setProperties in validateConfig.

Left the removed property labels in the resources bundle.