Changed addUserByName in LivelinkDocumentListIterator to ignore users (with a logged warning) when GetUserOrGroupByID fails to retrieve a user name for a given ID.
I can only recreate an error in this case by manually deleting a user
row from the KUAF table. Also, if this user is the DTree.CreatedBy
user, then the DTree row just disappears from the WebNodes view due to
the join against KUAF. But if the user is the DTree.UserID (that is,
the owner, rather than the creator), then you will get an error of
"Could not locate specified name. (402301)" from LAPI. (You will get a
similar error in the Livelink UI on the General Properties page.)