Fixed issue 20. The problem was that when getting the next batch of candidates, we ask for N rows from the database, but Livelink filters the results based on the user permissions. If a user does not have permission for an entire batch of candidates, we get no results at all back from Livelink, but there may be more in the database. We can't use permFirst. Among other issues, it doesn't funamentally change the equation of getting N candidates from the database and then filtering them by the user's permissions. Since we get no results at all, we can't even skip ahead.
There are several possible fixes: 

1. Use larger and larger batch sizes, such as 1000, 2000, 3000, etc.
We know each time that the previous batch size returned no results, so
we won't get more than 1000 candidates.

2. If we get no candidates, use the sysadminClient to look ahead for a
document to skip past.

3. Just use the sysadminClient to get candidates. We are only
retrieving DataID and ModifyDate, and the getResults will filter
everything through the user's permissions in Livelink. So no sensitive
data will be exposed by getting candidates as a system administrator.

This fix uses the third option. If we impersonate another user, we
create a separate client for the system administrator. The
getCandidatesSqlServer and getCandidatesOracle methods use the
sysadminClient instead of the client. I also changed the constructor
to be less dependent on ordering, by explicitly using the new
sysadminClient in isSqlServer and getStartCheckpoint, too.