Cleaned up URL validation.
* Added checks for some rare conditions that also indicate the URL was not
  valid. Most notably, the Sun Java System Web Server returns a 405
  (Method Not Allowed) instead of a 404 if you request a CGI program
  in an unknown directory.

* Added code to log the response code, but avoid logging exceptions
  that don't block URL validation. These might be confusing in the
  logs.

* Fixed issue 27, implicitly, by using a HEAD request and not reading
  the response body.

* Tried to avoid the "Server redirected too many times (20)" errors by
  not following redirects. We treat redirects as successes. I can't
  reproduce this error, so this is a bit of a stab in the dark.

Removed unneeded imports in CoreLivelinkConnectorTypeTest and its
subclass, LapiLivelinkConnectorTypeTest, as part of adding a test for
URL validation.

Added the response code and response message to the httpNotFound
error message, but only for English. They aren't critical, they appear
in the logs, and I didn't want to assume the format was portable
across all languages, although it's probably reasonable for most of
them.