Changed the query using excludedVolumeTypes to not include subvolumes. That is, the query is no longer hierarchical, and does not use DTreeAncestors. It just eliminates items where the OwnerID indicates a volume of one of the excluded types. I think this is what Livelink actually does.
The tests needed to change to expect the modified queries. The
DTreeAncestors validation in LivelinkConnector needed to change
because it is now possible to avoid needing DTreeAncestors (although
that is not the default configuration yet). The changes to
CoreLivelinkConnectorTypeTest were incidental: fixing the copyright
years, and the formatting of else clauses.