Fixed issue 31. This change is specific to the 1.1.x branch, and doesn't go on the trunk, where another fix was previously put in place.
Merged r280 changes into LivelinkTraversalManager, to limit the
skipping loop to two minutes. When that loop exits without finding any
documents, we return a specially crafted DocumentList to force the
call to checkpoint without waiting five minutes to resume the
traversal. An EmptyDocumentList contains one Document with a bogus
docid and action="delete". This is OK for QueryTraverser and
DocPusher, but ignored by the GSA.

Changed the OutOfMemory hack in LivelinkDocumentList (when all of the
calls to nextDocument throw an exception) to return an
EmptyDocumentList instead. This doesn't trigger the five minute wait,
and the OutOfMemoryError didn't seem to work with version 1.1 of the
connector manager anyway. The traversal didn't resume at all.

A      EmptyDocumentList.java
A      EmptyDocumentListTest.java
M      LivelinkDocumentList.java
M      LivelinkTraversalManager.java