Fixed issue 61: excludedLocationNodes only excluded children in the same parent volume as the excluded items, and not the excluded items themselves or children in the excluded item's volumes (such as projects).
Extracted the code handling includedLocationNodes in getIncluded,
which was correct, to a new getDescendants method, and used that in
getExcluded to handle excludedLocationNodes correctly.

This changed the excluded condition from "DataID not in (...)" to 
"not (DataID in (...) or DataID in (...))", which affected the tests.

M    projects/otex-core/source/java/com/google/enterprise/connector/otex/LivelinkTraversalManager.java
M    projects/otex-core/source/javatests/com/google/enterprise/connector/otex/LivelinkTraversalManagerTest.java