Followup to r440 to unify the setting of bean properties.
There are two types of pluggable properties in the connector. The
authenticationManager property uses beans, and the contentHandler
and genealogist properties use class names. (There's no good reason
for the difference.) The new authorizationManager and
publicContentAuthorizationManager properties use beans, but in a
slightly different way.

connectorDefaults.xml:

  <property name="authenticationManager" ref="authenticationManager"/>

  <property name="authorizationManager">
      <bean
class="com.google.enterprise.connector.otex.LivelinkAuthorizationManager"/>
  </property>

connectorInstance.xml:

  <bean id="authenticationManager"
class="com.google.enterprise.connector.otex.LivelinkAuthenticationManager"
        scope="prototype"/>

To override the property in connectorInstance.xml:

  <bean id="authenticationManager"
        class="com.example.OtherAuthenticationManager"
        scope="prototype"/>

  <property name="authorizationManager">
      <bean class="com.example.OtherAuthorizationManager"/>
  </property>

The problem with the old way was that it required the bean
definition to be in connectorInstance.xml, because if it was in
connectorDefaults.xml it could not be overriden (since
connectorDefaults.xml wins), thus defeating the point of the
property. Existing installations would not have the bean definition
in connectorInstance.xml, so either all customers would be forced to
add it, or we have to avoid the bean reference in
connectorDefaults.xml. An unresolved bean reference is an error,
rather than a null bean value.

This change switches authenticationManager to the new approach. This
means that customers who have overriden the bean definition will have
to override the property value when upgrading (either in addition to
or instead of the bean definition). I don't think there are any such
customers. Customers who have not overriden the bean definition will
have an extra bean definition for authenticationManager in their
connectorInstance.xml, but that shouldn't cause any problems.

Code review: http://codereview.appspot.com/4333049/

M    projects/otex-core/config/connectorDefaults.xml
M    projects/otex-core/config/connectorInstance.xml