Use H2 for all MockClient.ListNodes queries.
The only production code change was to switch from TO_DATE('...',
'YYYY-MM-DD HH24:MI:SS') to TIMESTAMP'...', because the latter works
with H2 (as well as Oracle 9i or later).

The in-memory database was changed from a per-connection database to a
shared in-memory database, to avoid having to share a single connection.
(Spring-instantiated connectors were the sticky point.) Instead of
relying on a fresh database for each test, we simply drop all the tables
in tearDown.

Many tests were OK with empty tables, some tests needed additional
values or columns added to the data. Only LivelinkAuthorizationManager
was tricky, because it used different query values and expected some of
them to return no results (easy) and multiple other values to return the
same database row (impossible without per-test method data). So I
crafted the row data and edited the expected values so that the various
possibilities were reflected in a single set of data.

I also generalized the correct-casing of the database columns in
MockClient, since the number of special cases grew a bit with the new
columns.

Code review: https://codereview.appspot.com/8696044/