Fix issue 6513973, part two: Unify SQL handling in other classes.
As with LivelinkTraversalManager, not everything has moved to the
resource bundles. The LivelinkConnector.autoDetectServtype and
LivelinkRetriever queries are not included.

I changed the authorization tests for hidden items to execute the
queries rather than inspect it, since I removed the getDocidQuery
method. I left the logic in getExcludedVolumeId alone, which means
that the same query is referenced twice, once with an explicit
choice(false) and once with choice(true). Combining them seemed too
convoluted, especially in conjunction with the explicit 0 return case
in that method.

Code review: https://codereview.appspot.com/8907044/