Fix issue 6032685, part one: Support OpenText Content Server 10. The column names passed to ListNodes are quoted.
This is actually an issue with recent updates of Livelink 9.7.1 as well.
The SQL injection protection had been, uh, changed, so that the column
names passed to ListNodes must be identifiers (with a few exceptions for
DISTINCT, *, T.*, and column [AS] alias). They are quoted as part of 
the constructed query. For example,

    ListNodes("<expression>", "<table>", { "col1", "col2" })

used to be turned into the following SQL query:

    select col1,col2 from <table> a where <expression>

and now it's:

    select "col1","col2" from <table> a where <expression>

The connector uses three basic kinds of queries that don't fit this new
requirement:

SELECT TOP n ...
SELECT function(...) ...
SELECT min(ModifyDate) ...

The fix is to move these expressions into the FROM or WHERE clauses. We
want to minimize the use of FROM expressions, though, because those have
to be whitelisted. So we really only want to move the second case,
non-aggregate expressions, into the FROM clause, because they are
modified column values and have to appear in the result set. The others
we can handle in the WHERE clause.

For TOP n queries, the modified query is something like this:

    select <original columns> 
    from <original table> 
    where <unique key> in (<original query>)

For the select expressions, the modified query is:

    select GoogleColumn as <original column>, <other columns>
    from (select b.*, <original expression> as GoogleColumn
          from <original table> b)
    where <original clause>

The min(ModifyDate) basically followed the TOP pattern, but the
non-uniqueness of the ModifyDate column made that a little inefficient
(we would select all the rows from DTree where the ModifyDate matched
the min value). So instead, the replacement is a small variation using
TOP 1 with an ascending ORDER BY on ModifyDate:

    select ModifyDate
    from DTree
    where DataID in (select TOP 1 DataID from DTree where <original clause>
                     order by ModifyDate)

MockClient.ListNodes applies the restriction on column names, so that
the unit tests will catch any invalid queries. The mocks are actually
slightly more restrictive than the real server, for a simpler
implementation and to enforce a consistent use of AS.

Notes:

* The select list code in LivelinkTraversalManager stayed there. There
  is some duplication with LivelinkRetriever. Since the latter is not
  production code, I was not highly motivated to refactor that
  duplication into efficient obscurity.

* Two untested methods were broken up so that the queries they ran could
  be tested directly using the production code (rather than executing
  the same queries separately in a test-only code path).

* The executeLimit method was removed. The price I paid was that the TOP
  expression is the same as the ROWNUM expression, that is to say, the
  last, even though TOP usually appears first in the text.

* A few numeric parameters that were handled as strings in
  LivelinkConnector are now ints. This was simply so that all numeric
  parameters are handled consistently with, e.g, {3,number,#} in the
  patterns, rather than special-casing the one query as {3}.

Code review: https://codereview.appspot.com/8957043/