Fix issue 6032685, part two: Support OpenText Content Server 10. All queries must select DataID or PermID.
This was always true for non-sysadmin users, and I think it was not
intended to be true for system administrators, but it is, at least in
OTCS 10 SP2 Update 10. It's a trivial change to three queries, and it
does not affect the execution plans (still no table scans, and no row
lookups that weren't already required).

I added this restriction to MockClient as well.

If you don't select DataID or PermID, all the rows in the result set
will be filtered out. I added a log message in a place where this
occurred silently.

Other changes:

* Logging all the selected query keys, to gather test coverage data
  during manual testing.

Code review: https://codereview.appspot.com/8958045/