Fix issue 9380415: DataSize over 2 GB triggers Livelink bug on Oracle and prevents traversal.
The mirror image of bug 4255719 in some ways, this one is ironically
caused by the fix for that one. When a DataSize value is 2 GB (2^31 or
2,147,483,647 bytes) or larger on Oracle, where DataSize is defined as
NUMBER(19), the case expression

    case when DataSize < 0 then 0 else DataSize end

leads to some kind of SQL error in the server (not the LAPI client).
This change adds a cast back to NUMBER(19) on Oracle:

    cast(case when DataSize < 0 then 0 else DataSize end as number(19))

This change affects a whitelisted view on Oracle, which is now:

(select b.*, cast(case when DataSize < 0 then 0 else DataSize end as number(19))
    as GoogleDataSize from WebNodes b)

Code review: https://codereview.appspot.com/9738055