Fix issue 9444414: HttpURLContentHandler should support timeouts.
Added properties for the connect and read timeouts, which are passed
through to the URLConnection. I added unit tests for the timeouts, to
ensure that an exception is thrown while reading the headers or body
of the response. I added a TODO to verify that these exceptions lead
to skipping the document in both cases, which I verified manually.

Other changes:

* Removed the unused connector field.

* Made the urlBase and urlPath fields private.

Code review: https://codereview.appspot.com/10309043