Fix issue 9580681: Handle cookie expiration in HttpURLContentHandler.
We use the LLCookie for authentication, but we only get the cookie
once at startup. By default, the cookie expires after 8 days, which is
a long time, but the connector can easily be running longer than that.
Customers can also configure the expiration of the cookie in their
Livelink servers.

HttpURLContentHandler is now a RefreshableContentHandler, which adds a
refresh method that is called once per batch. The implementation just
gets a new cookie value.

The MockClient's GetCookieInfo now returns a unique cookie value on
each call so we can test that a new cookie was requested. Added
EasyMock to the tests, using the connector manager's JAR files.

Code review: https://codereview.appspot.com/10671043/