Workaround a LAPI bug in the bad candidates set retry.
Livelink only stores timestamps to the nearest second, but LAPI 9.7
and earlier constructs a Date object that includes milliseconds, which
are taken from the current time. So we need to avoid using the
milliseconds in the Livelink date.

The bug is fixed in LAPI 9.7.1, so I updated all the comments about
this bug in the code accordingly.

Code review: https://codereview.appspot.com/13320043/