Fix issue 11484159: Oracle minModifyDate query is excessively slow in a customer environment.
For some reason the existing query, intended to scan the DTree_Google
index with a join against DTreeAncestors, was doing a table lookup by
rowid for millions of rows in DTree. Switched to a customer suggested
query that I had previously rejected because the execution plan is
much more complicated, but it does have fewer logical reads even in
our environment (which does not have the table lookups). It also
happens to have been the original query used here before ListNodes
views required whitelisting.

The interim query, introduced with support for whitelisting in r321,
modified the meaning of the query and rendered it useless as a
validation check on DTreeAncestors. It was now useful only to set the
start date. I do not think that change was completely intentional,
since none of the method comments or method name were updated
accordingly. Returning to the join restores the original semantics: at
least one of the Items to Index or their descendants must appear in
DTreeAncestors.

I considered switching the SQL Server query in a similar fashion, but
the current plan has the lowest cost and very similar execution plans
to the other variations, I have no compelling reason to change it, and
there's a risk that a modified query would introduce problems.

Added a test in LivelinkConnectorTest that explicitly checks the
results of this query.

Updated test data in LivelinkTraversalManagerTest that didn't meet the
stronger original requirements for membership in DTreeAncestors.

Other changes:

* Switched other tests in LivelinkConnectorTest to use the
  LivelinkDateFormat object I added, to avoid expected date strings
  with hard-coded time zones

* Updated comments in LivelinkConnector about not selecting DataID.
  OTCS 10 always requires DataID to be selected, and that was changed
  in r549.

Code review: https://codereview.appspot.com/21820043/