Test cases for Livelink group feed using Adaptor library.
New test property tests.httpServerPort is added to 
"google-enterprise-connector-otex.properties.sample". 
This is the port to be used for starting SimpleHttpServer. 
SimpleHttpServer is used to server GSA version to Adaptor. 
GSA uses port 80. httpServerPort is the port that is used for 
redirecting traffic from port 80 on test machine. 

On Mac, for example, use "sudo ipfw add 100 fwd 127.0.0.1,8080 
tcp from any to any 80 in" to redirect traffic from port 80 to 
port 8080.

Code review: https://codereview.appspot.com/131720043/