Fix inadequate locking in GroupLister.
Resetting the connector traversal would start an extra adaptor
instance, and leave the non-daemon threads of the embedded HttpServer
running. This prevented Tomcat from shutting down.

The calls to start and shutdown a Lister may be called asynchronously,
due to a race condition in the CM, and multiple times, by design. The
CM may call shutdown-start-shutdown-start [sic] in quick succession,
and start executes in a worker thread while shutdown executes in the
main thread, so the lister instance may see this as
shutdown-shutdown-start-start. This is probably a CM bug, but may be
difficult to fix there since Lister.start is designed not to exit until
shutdown is called. The multiple pairs of calls can cause problems if
the Lister implementation holds state in fields, which can be
overwritten by multiple calls to start.

This change creates uses a lock for both start and shutdown, so that
both may not be running simultaneously (with some necessary
exceptions), and so that calls to start and stop must alternate. We
block at the top of start until the adaptor is stopped, and at the top
of shutdown until the adaptor is started.

To replace the existing the Semaphore for shutdown, we also use the
same Condition in start to block after starting the adaptor until
shutdown is called.

Condition and Lock are like the builtin wait/notify/notifyAll and
sychronized features, except that multiple conditions can be
associated with the same lock. I use that as a small optimization so
that start only needs to notify a thread in shutdown, rather than all
threads, and vice versa.

Timed waits are used for the mutual exclusion waits, since these waits
are only needed to avoid race conditions. If the waits timeout a
RepositoryException is thrown and the Lister is assumed in both calls
(start and shutdown) to have exited. Similarly, if a waiting thread is
interrupted (and throws an InterruptedException), the Lister is assumed
in both calls to have exited. The wait that is blocking start from
exiting is not timed or interruptible, since it should run indefinitely
by design (until the Lister is shutdown).

Other changes:

* Add a configurable dashboardPort for testability. The production code
  still uses 0, and some of the tests explicitly use the production
  code path when they don't care about the dashboard port.

* Convert GroupListerTest to JUnit 4. I was using
  @Test(expected=SomeException.class) briefly, though in the end the
  change was superfluous.

* Avoid use of getBytes() in SimpleHttpServer and use US_ASCII.

* Add adaptor package logging to the tests. 

Code review: https://codereview.appspot.com/172350043/